requirements.txt
streamlit
yfinance
ta
plotly
pandas
python-dotenv
requests
openpyxl
scikit-learn
numpy
scipy
textblob
vaderSentiment
